mchanged implementation for passing data, added new requests
