+++ servers/data.json
@@ -1,0 +1,1 @@
+"{ \"name\":\"John\", \"age\":30, \"city\":\"New York\"}"
