+++ servers/index.html
@@ -1,0 +1,10 @@
+<!DOCTYPE html> 
+<html lang="en"> 
+  <head> 
+    <title>Button in Node.js</title> 
+  </head> 
+  <body>
+    <form action="http://localhost:3000/" method="GET"> 
+      <button>Call Server</button> 
+    </form> 
+</html>
