+++ servers/servers.txt
@@ -5,3 +5,13 @@ Steps to running servers:
 4) "node server.js" (run second server)
 5) Go to "localhost:3000" then "localhost:4000" (expect to see nothing)
 6) return to terminals to see server interaction via console logs
+
+/*
+  Unity server -> makes request to simulator (passes user data json)
+  then simulator passes back json (different file, visual version of data) 
+  to unity game to start game.
+
+  testing: python -> outputs json to a file
+  get json data file then return it (response, perhaps 
+    listen for post/get request)
+*/
