+++ servers/test.py
@@ -1,4 +1,6 @@
-a = 10
-print(a)
+import json
 
-# make this print json
+x =  '{ "name":"John", "age":30, "city":"New York"}'
+
+with open('data.json', 'w') as outfile:
+    json.dump(x, outfile)
