+++ servers/views/index.ejs
@@ -1,0 +1,13 @@
+<!DOCTYPE html> 
+<html lang="en"> 
+<head> 
+<meta charset="UTF-8"> 
+<meta name="viewport" content="width=device-width, initial-scale=1.0"> 
+<meta http-equiv="X-UA-Compatible" content="ie=edge"> 
+<title>Document</title> 
+</head> 
+<body> 
+<button id="btn" onclick=>Button Here </button> 
+<script src="website.js"></script> 
+</body> 
+</html>
